Changes from ulong to long in mpfr_vec, mpfr_poly, nmod_vec, and nmod_poly

+++ nmod_poly/mul.c
@@ -36,8 +36,8 @@ void nmod_poly_mul(nmod_poly_t res, nmod_poly_t poly1, nmod_poly_t poly2)
       return;
    }
    
-   ulong bits = FLINT_BITS - poly1->mod.norm;
-   ulong bits2 = FLINT_BIT_COUNT(FLINT_MAX(poly1->length, poly2->length));
+   long bits = FLINT_BITS - (long) poly1->mod.norm;
+   long bits2 = FLINT_BIT_COUNT(FLINT_MAX(poly1->length, poly2->length));
 
    if (2*bits + bits2 <= FLINT_BITS)
       nmod_poly_mul_classical(res, poly1, poly2);
